+++ packages/international-phone-input/src/components/country-select/index.ts
@@ -1,0 +1,1 @@
+export { CountrySelect, SharedCountrySelectProps } from './Component';