+++ packages/shared/src/index.ts
@@ -5,3 +5,4 @@ export * from './easingFns';
 export * from './inputUtils';
 export * from './fnUtils';
 export * from './isMobile';
+export * from './maskUtils';