+++ packages/custom-button/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## 3.3.5
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   button@9.0.5
+
 ## 3.3.4
 
 ### Patch Changes