+++ packages/gallery/CHANGELOG.md
@@ -1,5 +1,20 @@
 # Change Log
 
+## 5.3.11
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   tooltip@7.1.6
+    -   icon-button@6.2.5
+
+## 5.3.10
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   tooltip@7.1.5
+
 ## 5.3.9
 
 ### Patch Changes